Added Dutch Constitution
Added the Constitution (Grondwet) of the Kingdom of the Netherlands, revision of the year A.D. 2018, in their respective languages, as publicly listed on the website of the Rijksoverheid (Dutch Government), Dutch, English, French, German and Spanish.

Warning: This document is a legal document of the Dutch parliamentary democracy and rule of law, it stands above the Microsoft and GitHub Terms of Service, may not be modified or re used without permission of the Dutch States-General, modification, sharing modified veriants, and other forms of misuse, are deemed criminal acts under Dutch and European law, and shall be subject to Dutch and European courts. These documents are made available by me purely for informational purposes, as such it is allowed to list them. If you make any unauthorized edits and publish these, I, GitHub and Microsoft will NOT be held responsible.